refactor(path): 길찾기 API의 폴더 구조 이동 - 기존 domain 폴더에서 interfaces로 이동
